Modularize create_ggnn_input.py
